+++ packages/contracts/deploy/new/40_finalize-management-dao/30_install-multisig-on-management-dao.ts
@@ -6,6 +6,11 @@ import {
   PluginSetupProcessor__factory,
 } from '../../../typechain';
 import {InstallationPreparedEvent} from '../../../typechain/PluginSetupProcessor';
+import {
+  isLocal,
+  managementDaoMultisigApproversEnv,
+  managementDaoMultisigMinApprovalsEnv,
+} from '../../../utils/environment';
 import {hashHelpers} from '../../../utils/psp';
 import {checkPermission, getContractAddress} from '../../helpers';
 import {findEvent} from '@aragon/osx-commons-sdk';
@@ -18,7 +23,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const {ethers, network} = hre;
   const [deployer] = await ethers.getSigners();
 
-  if (network.name !== 'localhost' && network.name !== 'hardhat') {
+  if (!isLocal(network)) {
     if (
       !('MANAGEMENT_DAO_MULTISIG_LISTEDONLY' in process.env) ||
       !('MANAGEMENT_DAO_MULTISIG_MINAPPROVALS' in process.env) ||
@@ -28,11 +33,9 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     }
   }
 
-  const approvers = process.env.MANAGEMENT_DAO_MULTISIG_APPROVERS?.split(
-    ','
-  ) || [deployer.address];
+  const approvers = managementDaoMultisigApproversEnv(network).split(',');
   const minApprovals = parseInt(
-    process.env.MANAGEMENT_DAO_MULTISIG_MINAPPROVALS || '1'
+    managementDaoMultisigMinApprovalsEnv(hre.network)
   );
   // In case `MANAGEMENT_DAO_MULTISIG_LISTEDONLY` not present in .env
   // which applies only hardhat/localhost, use `true` setting for extra safety for tests.