+++ packages/contracts/deploy/verification/99_conclude/10_verify-contracts.ts
@@ -1,3 +1,4 @@
+import {isLocal} from '../../../utils/environment';
 import {verifyContract} from '../../../utils/etherscan';
 import {DeployFunction} from 'hardhat-deploy/types';
 import {HardhatRuntimeEnvironment} from 'hardhat/types';
@@ -29,8 +30,4 @@ export default func;
 func.tags = ['New', 'Verify', 'VerifyContracts'];
 func.runAtTheEnd = true;
 func.skip = (hre: HardhatRuntimeEnvironment) =>
-  Promise.resolve(
-    hre.network.name === 'localhost' ||
-      hre.network.name === 'hardhat' ||
-      hre.network.name === 'coverage'
-  );
+  Promise.resolve(isLocal(hre.network));